TIP-860: Move MeasureFamily related api tests and Controller
