+++ src/Akeneo/Tool/Bundle/MeasureBundle/tests/EndToEnd/ExternalApi/ListMeasureFamilyIntegration.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\MeasureFamily;
+namespace Akeneo\Tool\Bundle\MeasureBundle\tests\EndToEnd\ExternalApi;
 
 use Akeneo\Tool\Bundle\ApiBundle\tests\integration\ApiTestCase;
 use Symfony\Component\HttpFoundation\Response;
